backend_ai/requirements.txt
# ─── Web Framework ────────────────────────────
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic==2.11.3
pydantic-settings==2.9.1

# ─── HTTP Client ──────────────────────────────
httpx==0.28.1

# ─── LLM Providers ───────────────────────────
openai==1.82.0
anthropic==0.52.0

# ─── Web Scraping / Parsing ──────────────────
beautifulsoup4==4.13.4
trafilatura==2.0.0
lxml==5.4.0

# ─── PDF Parsing ─────────────────────────────
pypdf==5.6.0

# ─── Search Providers ────────────────────────
# tavily-python is optional; install if using Tavily
# tavily-python==0.5.0

# ─── Utilities ───────────────────────────────
python-dateutil>=2.9.0
python-dotenv==1.1.0
